+++ src/components/ids-data-grid/demos/editable.html
@@ -47,7 +47,8 @@
           row-selection="multiple"
           label="Books"
           editable="true"
-          edit-next-on-enter-press="true">
+          edit-next-on-enter-press="true"
+          add-new-at-end="true">
         </ids-data-grid>
       </ids-layout-grid-cell>
     </ids-layout-grid>
